New generic method JSON->CLASS. Take "json-y" input and store in specified class.
Several specializers are defined. One :around method makes sure the
class registry for FLUID-OBJECTs is empty and tells CL-JSON to look in
:NISP.GITHUB for classes.

The :after method specializing on when OBJECT is a stream makes sure the
stream is closed.

One primary method is responsible for returning classes, either FLUID or
STANDARD-CLASS.