Add NISP:%GLOBAL
I wrote this, then realized that most of the time its not going to matter,
but these are there in case I need them and at least lets me tell what
machine I'm on. However I know its a bad idea to use anything in this
package in something meant to be portable.