Add TREE-SYMBOL-STRING type.
This lets me tell if something is a valid "thing" to be converted to a
command string function. All commands are seperated by spaces. (and
possibly other things later) so we need to make sure we don't intern
symbols with these charactors or there is absolutly no chance these new
functions will ever get called.