Define TREE-GENERIC-DIRECT-NODES
This gets superclassed by both TREE-GENERIC-FUNCTION and TREE-SPECIALIZER
which now means methods that want to operate on the hash tables contained
that these two classes contained now can use the same interface and save
on code duplication as we don't have to write a case for
TREE-GENERIC-FUNCTION and a case for TREE-SPECIALIZER when we don't care
which of the two we have, so long as it has that hash table.

This commit will break a ton of test cases likely as the reader methods
TREE-GENERIC-FUNCTION-ROOT-NODES and TREE-SPECIALIZER-DIRECT-NODES have
been removed. I've chosen to do this now rather then later as nothing in
the code set being changed right now actually worked to start with, so
breaking something broken already (and covered by test cases for what does
work) does little to no harm.