Do not define INTERN-TREE-SPECIALIZER as a method but as a function
Its far more elegant (in my eyes) for us to recursively call an external
function (that I can trace!) then to try to do the same thing in a method
using LABELS as we don't really have any need to specialize on type input,
and I (right now) don't see a need to have before or after methods on
specializer interning.

However translating this to a generic function using an internal LABELS
for the recursion part is not totally out of the question, but I don't
want to do it until I'm a little more sure that what we are doing is
correct behavior.