TREE-GENERIC-FUNCTION uses COMMON-LISP:STANDARD-GENERIC-FUNCTION now.
We do this instead of using closer-mop's version is because we do not use
any of the features of closer-mop's version, and that version has an &rest
parameter in it, making any attempts to modify and extend
MAKE-METHOD-LAMBDA kill any effeciency ideas.

In summary, with it, I was getting roughly 330 method calls per second
where the method simply did (list arg1 arg2), without it, I got native
sbcl speed. This is documented in tree-funcallable-class.org if anyone is
curious.