The bot will now use ROUTE instead of ROUTE-COMMAND.
As of this commit the IRC bot is now part of a larger routing system that
will hopefully start to show some real gains soon from all the
research/theory as defining new handlers is now near trivial.

This is going to take a few more days before commands can be added with
any sort of speed (show a real gain for the investment) as the current
command is defined to shake out some bugs first.