ENSURE-NETWORK-NODE-SYMBOL: Don't use MAKE-KEYWORD
We call (intern <the string upcased> +NETWORK-TREE-SYMBOLS-PACKAGE+)

It calls (intern (string <the string>) :keyword)

We get roughly 20% speed improvement by passing the package in instead of
making intern look up the package. This is with no special declares or
speed settings, and measured over 1,000,000 calls.

This improvement matters in a function that gets called often during the
course of network-node lookup