Nodes now stored with string keys.
This is way faster then interning and then doing lookup using 'eq hash key
lookups, equalp is far faster for all the cases I've tested where the
input is not interned or found yet. The killer is converting arbitrary
input to the symbol representation, There has to be a better way then
equalp on strings for the keys, especially when the input is not
fundementally a string to start with... but for now this is what I'm doing
until something better pops up.