1000 times faster then prior working version!
This commit marks the version of the library that works in the same way as
the older stable version, but we are now literally 1000 times faster by
relying on native SBCL MOP where possible instead of the wrapper
CLOSER-MOP.

A test sequence that took 1 second to do 330 calls now can do those 330
calls in 0.001 seconds.

This commit adds in fundemental changes to MAKE-METHOD-LAMBDA and mostly
reverts to what we had before with respect to
COMPUTE-DISCRIMINATING-FUNCTION. Not all work is done with respect to
functionality but this commit marks a real working version.