Move SHORTURL-IS.GD to the top of the file
I aim to move this stuff somewhere else other then bot.lisp, but some of
this stuff is not really worth its own file, and is.gd begs for its own
package/library as it is an independent webapi library, in all of one
function! But that is way overkill of course ;).