Found some cool emulators and just a few assembler instruction notes.
The emulators were found with the help of debian GNU/Linux debtags. They
are all listed in the hardware::emulator debtag. I just read the
descriptions and took notes on the ones I found to be of interest to
me (quite a few!) and I'll start exploring these with an eye to finding
some interesting instruction sets.